+++ api-clinica/app/Http/Controllers/Patient/PatientController.php
@@ -8,9 +8,12 @@ use App\Models\Patient\Patient;
 use Illuminate\Support\Facades\DB;
 use App\Http\Controllers\Controller;
 use App\Models\Patient\PatientPerson;
+use Illuminate\Support\Facades\Redis;
+use App\Models\Appointment\Appointment;
 use Illuminate\Support\Facades\Storage;
 use App\Http\Resources\Patient\PatientResource;
 use App\Http\Resources\Patient\PatientCollection;
+use App\Http\Resources\Appointment\AppointmentCollection;
 
 class PatientController extends Controller
 {
@@ -19,6 +22,7 @@ class PatientController extends Controller
      */
     public function index(Request $request)
     {
+        $this->authorize('viewAny',Patient::class);
         $search = $request->search;
 
         $patients = Patient::where(DB::raw("CONCAT(patients.name,' ',IFNULL(patients.surname,''),' ',patients.email)"),"like","%".$search."%")
@@ -31,11 +35,72 @@ class PatientController extends Controller
         ]);
     }
 
+    public function profile($id)
+    {
+        $this->authorize('profile',Patient::class);
+        $cachedRecord = Redis::get('profile_patient_#'.$id);
+        $data_patient = [];
+        if(isset($cachedRecord)) {
+            $data_patient = json_decode($cachedRecord, FALSE);
+        }else{
+            $patient = Patient::FindOrFail($id);
+    
+            $num_appointment = Appointment::where("patient_id",$id)->count();
+            $money_of_appointments = Appointment::where("patient_id",$id)->sum("amount");
+            $num_appointment_pendings = Appointment::where("patient_id",$id)->where("status",1)->count();
+            
+            $appointment_pendings = Appointment::where("patient_id",$id)->where("status",1)->get();
+            $appointments = Appointment::where("patient_id",$id)->get();
+            $data_patient = [
+                "num_appointment" => $num_appointment,
+                "money_of_appointments" => $money_of_appointments,
+                "num_appointment_pendings" => $num_appointment_pendings,
+                "patient" => PatientResource::make($patient),
+                "appointment_pendings" => AppointmentCollection::make($appointment_pendings),
+                "appointments" => $appointments->map(function($appointment){
+                    return[
+                        "id" => $appointment->id,
+                        "patient" =>[
+                            "id" => $appointment->patient->id,
+                            "full_name" => $appointment->patient->name . ' ' .$appointment->patient->surname,
+                            "avatar" => $appointment->patient->avatar ? env("APP_URL")."storage/".$appointment->patient->avatar : NULL,
+                        ],
+                        "doctor" => [
+                            "id" => $appointment->doctor->id,
+                            "full_name" => $appointment->doctor->name . ' ' .$appointment->doctor->surname,
+                            "avatar" => $appointment->doctor->avatar ? env("APP_URL")."storage/".$appointment->doctor->avatar : NULL,
+                        ],
+                        "date_appointment"  => $appointment->date_appointment,
+                        "date_appointment_format" => Carbon::parse($appointment->date_appointment)->format("d M Y"),
+                        "format_hour_start" => Carbon::parse(date("Y-m-d").' '.$appointment->doctor_schedule_join_hour->doctor_schedule_hour->hour_start)->format("h:i A"),
+                        "format_hour_end"=> Carbon::parse(date("Y-m-d").' '.$appointment->doctor_schedule_join_hour->doctor_schedule_hour->hour_end)->format("h:i A"),
+                        "appointment_attention" => $appointment->attention ? [
+                            "id" => $appointment->attention->id, 
+                            "description" => $appointment->attention->description,
+                            "receta_medica" => $appointment->attention->receta_medica ? json_decode($appointment->attention->receta_medica) : [],
+                            "created_at" => $appointment->attention->created_at->format("Y-m-d h:i A"),
+                            ] : NULL,
+                            "amount"  => $appointment->amount,
+                            "status" => $appointment->status,
+                            "status_pay"  => $appointment->status_pay,
+                    ];
+                }),
+            ];
+            Redis::set('profile_patient_#'.$id, json_encode($data_patient),'EX', 3600);
+        }
+
+
+        return response()->json($data_patient);
+    }
+
+
+
     /**
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
     {
+        $this->authorize('create',Patient::class);
         $patient_is_valid = Patient::where("n_document",$request->n_document)->first();
 
         if($patient_is_valid){
@@ -72,6 +137,7 @@ class PatientController extends Controller
      */
     public function show(string $id)
     {
+        $this->authorize('view',Patient::class);
         $patient = Patient::FindOrFail($id);
 
         return response()->json([
@@ -84,6 +150,7 @@ class PatientController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $this->authorize('update',Patient::class);
         $patient_is_valid = Patient::where("id","<>",$id)->where("n_document",$request->n_document)->first();
 
         if($patient_is_valid){
@@ -126,9 +193,14 @@ class PatientController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize('delete',Patient::class);
         $patient = Patient::findOrFail($id);
         if($patient->avatar){
             Storage::delete($patient->avatar);
+        }
+        $cachedRecord = Redis::get('profile_patient_#'.$id);
+        if(isset($cachedRecord)) {
+            Redis::del('profile_patient_#'.$id);  
         }
         $patient->delete();
         return response()->json([
